Support limited LibZMQ functionality needed for Push Jobs Client
1) Add in support for the rake compiler extension for testing
2) Update Gemfile/gemspec extensions to modern patterns
3) Only implement c functions we directly need in push jobs
