+++ lib/rbzmq.rb
@@ -1,0 +1,1 @@
+require 'rbzmq/ext'
