+++ lib/rbzmq/version.rb
@@ -1,0 +1,3 @@
+module ZMQ
+  VERSION = "3.0.0"
+end
